+++ src/components/ProjectGallery/ProjectGallery.tsx
@@ -1,49 +1,18 @@
 import './ProjectGallery.css'
-import img1 from '@statics/images/correlation/correlation1.png';
-import img2 from '@statics/images/correlation/correlation2.png';
-import img3 from '@statics/images/correlation/correlation3.png';
-import img4 from '@statics/images/correlation/correlation4.png';
-
 import { BsArrowRightCircle, BsArrowLeftCircle } from "react-icons/bs";
-
 import { useState, useEffect, useRef } from 'react'
+import { SlideShowOBJ } from '../../pages/Project/types';
+import { NumberLocale } from 'yup/lib/locale';
 
-
-interface SlideShowOBJ {
-    img: string;
-    title: string;
-    description: string;
+interface PropsOBJ {
+    itemArray: Array<SlideShowOBJ>;
+    displayNumber: number;
+    compTitle: string;
 }
 
-const dummyList: SlideShowOBJ[] = [
-    {
-        img: img1,
-        title: 'Figure 1.1',
-        description: "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum."
-    },
-    {
-        img: img2,
-        title: 'Figure 1.2',
-        description: "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum."
-    },
-    {
-        img: img3,
-        title: 'Figure 1.3',
-        description: "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum."
-    },
-    {
-        img: img4,
-        title: 'Figure 1.4',
-        description: "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum."
-    },
-    {
-        img: img1,
-        title: 'Figure 1.5',
-        description: "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum."
-    }
-]
+const ProjectGallery = (props:PropsOBJ) => {
 
-const ProjectGallery = () => {
+    const { itemArray, displayNumber, compTitle } = props;
 
     // This defines the index of which element is being displayed within the gallery at the moment
     const [galleryIndex, setGalleryIndex] = useState<number>(0);
@@ -62,7 +31,7 @@ const ProjectGallery = () => {
     useEffect(() => {
         resetTimeout()
         timeoutRef.current = setTimeout(
-            () => setGalleryIndex((prev) => prev === dummyList.length - 1 ? 0 : prev + 1), 4000
+            () => setGalleryIndex((prev) => prev === itemArray.length - 1 ? 0 : prev + 1), 4000
         )
         return () => {
             resetTimeout()
@@ -81,16 +50,16 @@ const ProjectGallery = () => {
                     }
                     <div>
                         <h1 id='container-title-h1'>
-                            02
+                            {displayNumber}
                         </h1>
                         <h2 id='container-title-h2'>
-                            Gallery
+                            {compTitle}
                         </h2>
                     </div>
                 </div>
                 <div id='gallery-container'>
                     < BsArrowLeftCircle className='gallery-buttons' color='white' size={"2.5rem"} onClick={() => {
-                        galleryIndex === 0 ? setGalleryIndex(dummyList.length - 1) : setGalleryIndex(galleryIndex - 1)
+                        galleryIndex === 0 ? setGalleryIndex(itemArray.length - 1) : setGalleryIndex(galleryIndex - 1)
                     }} />
                     <div className="gallery-box">
                         <div className='slideshowSlider' style={{ transform: `translate3d(${-galleryIndex * 100.5}%, 0, 0)` }}>
@@ -98,7 +67,7 @@ const ProjectGallery = () => {
                                 /* 
                                     Theses displayed properties should also be from props
                                 */
-                                dummyList.map((obj, index) => (
+                                    itemArray.map((obj, index) => (
 
                                     <div className='box-content' key={index}>
                                         <div className='image-container'>
@@ -112,7 +81,7 @@ const ProjectGallery = () => {
                         </div>
                         <div className='slideshowDots'>
                             {
-                                dummyList.map((_, idx) => (
+                                itemArray.map((_, idx) => (
                                     <div
                                         key={idx}
                                         className={`slideshowDot${galleryIndex === idx ? " active" : ""}`}
@@ -125,7 +94,7 @@ const ProjectGallery = () => {
                         </div>
                     </div>
                     < BsArrowRightCircle className='gallery-buttons' color='white' size={"2.5rem"} onClick={() => {
-                        galleryIndex === dummyList.length - 1 ? setGalleryIndex(0) : setGalleryIndex(galleryIndex + 1)
+                        galleryIndex === itemArray.length - 1 ? setGalleryIndex(0) : setGalleryIndex(galleryIndex + 1)
                     }} />
                 </div>
             </div>
