+++ src/pages/Project/types.tsx
@@ -1,0 +1,5 @@
+export interface SlideShowOBJ {
+    img: string;
+    title: string;
+    description: string;
+}
